+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * Implement changing bucket settings, [PR-9](https://github.com/reduct-storage/web-console/pull/9)
 * Show error in CreateOrUpdate form if don't get settings, [PR-10](https://github.com/reduct-storage/web-console/pull/10)
 
+**Bugs**:
+
+* Fix bigint conversions in bucket settings, [PR-11](https://github.com/reduct-storage/web-console/pull/11)
+
 **Other**:
 
 * Update reduct-js to 0.4.0, [PR-6](https://github.com/reduct-storage/web-console/pull/6)